link guild subcommand
